pyproject.toml
[tool.poetry]
name = "poetry-config"
version = "0.1.0"
description = "API RESTFUL with Django Rest Framework"
authors = ["Nidal Chateur <nidalchateur@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
djangorestframework = "^3.14.0"
requests = "^2.31.0"
djangorestframework-simplejwt = "^5.3.0"
pillow = "^10.0.1"
flake8 = "^6.1.0"
flake8-html = "^0.4.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
